Complete exporters.py modularization
- Created map_locations/exporters.py with 5 export functions
- Reduced core.py from 1,755 to 1,213 lines (-542 lines)
- Updated core.py to import from exporters module
- All backward compatibility maintained
- Testing: JSON, CSV, GeoJSON, and KML exports work correctly

Progress: 83% Complete (5/6 modules)
ALL PRIORITY 2 CONTENT MODULES COMPLETE
Next: map_generators.py (Priority 3: High Risk - largest remaining module)
